backend/requirements.txt
fastapi
uvicorn
python-pptx
motor
pydantic
python-dotenv
meaningless
google-generativeai
deep-translator
fuzzywuzzy
python-Levenshtein
pandas
lxml
